requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.9
django-cors-headers==4.3.1
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
drf-yasg==1.21.7
filetype==1.2.0
flake8==7.0.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
setuptools==69.2.0